+++ src/entities/form/constants/selectOptionData.tsx
