+++ src/entities/form/index.tsx
@@ -9,3 +9,4 @@ export { default as RequiredToggle } from './ui/RequiredToggle';
 export { default as CreateFormButton } from './ui/CreateFormButton';
 export { default as CheckBox } from './ui/CheckBox';
 export { default as PrivacyConsentForm } from './ui/PrivacyConsentForm';
+export { selectOptionData } from './constants/selectOptionData';
