+++ src/features/form/common/index.ts
@@ -1,0 +1,1 @@
+export { default as FormEditor } from './ui/FormEditor';
