+++ src/features/form/create/model/useCreateApplicationForm.ts
@@ -17,7 +17,7 @@ export const useCreateApplicationForm = (
       createApplicationForm({ data: formattedData, id }),
     onSuccess: () => {
       toast.success('박람회 등록 폼이 생성되었습니다.');
-      router.push(`/expo-detail/${id}`);
+      router.push(`/exhibition/detail//${id}`);
       queryClient.resetQueries({
         queryKey: ['createApplicationForm', id, type],
       });
