+++ src/features/form/create/model/useCreateSurveyForm.ts
@@ -17,7 +17,7 @@ export const useCreateSurveyForm = (
       createSurveyForm({ data: formattedData, id }),
     onSuccess: () => {
       toast.success('만족도 조사 폼이 생성되었습니다.');
-      router.push(`/expo-detail/${id}`);
+      router.push(`/exhibition/detail//${id}`);
       queryClient.resetQueries({
         queryKey: ['createSurveyForm', id, type],
       });
